Move function from file.py to logcollector.py
